Re-work the price fetch mechanism to avoid extra calls to AlphaVantage which led to rate limit issues

+++ stockdash-backend/src/main/java/com/temadison/stockdash/backend/api/dto/PositionValueDto.java
@@ -4,7 +4,7 @@ import java.math.BigDecimal;
 
 public record PositionValueDto(
         String symbol,
-        long quantity,
+        BigDecimal quantity,
         BigDecimal currentPrice,
         BigDecimal marketValue
 ) {
